Working Database - Able to add new (positive value only) transactions

+++ app/src/main/java/com/lewiswilson/minimalistsavingstracker/RecyclerItem.java
@@ -1,17 +1,15 @@
 package com.lewiswilson.minimalistsavingstracker;
 
-import java.math.BigDecimal;
-
 public class RecyclerItem { //used to fill the recyclerview
-    private BigDecimal mAmount;
+    private int mAmount;
     private String mReference;
 
-    public RecyclerItem(BigDecimal amount, String reference) { //define an item
+    public RecyclerItem(int amount, String reference) { //define an item
          mAmount = amount;
          mReference = reference;
     }
 
-    public BigDecimal getAmount() { //getting the amount value
+    public int getAmount() { //getting the amount value
         return mAmount;
     }
